+++ src/traits/mod.rs
@@ -4,4 +4,6 @@ mod source_err;
 
 pub use contextual::ErrorWith;
 pub use conversion::{ConvErr, ConvStructError, ToStructError};
-pub use source_err::{raw_source, RawSource, RawStdError, SourceErr};
+pub use source_err::{
+    any_err, raw_source, AnyErr, RawSource, RawStdError, SourceErr, SourceRawErr,
+};